% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combo_functions.R
\name{bind_nmfs_dfo_tows}
\alias{bind_nmfs_dfo_tows}
\title{Bind NMFS and DFO tow data}
\usage{
bind_nmfs_dfo_tows(nmfs_tows, dfo_tows, out_dir)
}
\arguments{
\item{nmfs_tows}{= NMFS tows dataframe, created by `nmfs_get_tows()`}

\item{dfo_tows}{= DFO tows dataframe, created by `dfo_get_tows()`}

\item{out_dir}{= Directory to save the combined dataframe as an .rds file}
}
\value{
A datafame with information of all unique tows. This file is also saved in out_dir.
}
\description{
A short function that will row bind a NMFS tows dataframe and DFO tows dataframe to create one dataframe that has all tows combined, with one row per unique tow.
}
