% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfo_functions.R
\name{dfo_get_tows}
\alias{dfo_get_tows}
\title{DFO tow data}
\usage{
dfo_get_tows(dfo_GSINF, dfo_GSMISSIONS, out_dir)
}
\arguments{
\item{out_dir}{Directory to save the cleaned dataframe as an .rds file}
}
\value{
A datafame with information of all unique DFO tows and matching column names to NMFS data. This file is also saved in out_dir.
}
\description{
This function uses the `dfo_GSINF` and `dfo_GSMISSIONS` databases to create a dataframe that has one row for every unique tow, with column names that match those readily available from the NOAA NMFS surveys.
}
