% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate_functions.R
\name{dynamic_2d_extract}
\alias{dynamic_2d_extract}
\title{Extract dynamic covariates at point locations}
\usage{
dynamic_2d_extract(
  rast_ts_stack,
  stack_name,
  t_summ,
  t_position,
  sf_points,
  date_col_name,
  df_sf
)
}
\arguments{
\item{rast_ts_stack}{= A raster stack of one covariate variable, where each layer represents a different time step}

\item{stack_name}{= The column name to use for the extracted value for the dynamic covariate}

\item{t_summ}{= A numeric value or character string that indicates what temporal resolution should be used in summarizing the covariate values. If numeric, the function will simply summarize the values in the raster stack from the matching period back `t_summ` numeric time steps. For example, if the `rast_ts_stack` provided daily values and t_summ = 90, the function would calculate a 90 day average, where the 90-day window would either be leading up to and including the day of observation, saddled around the observation, or include the day of the observation and 89 days into the future. If a character string, should be one of "daily", monthly", "seasonal", or "annual". These options are built into the function to provide a bit easier specification to quickly calculate the monthly/seasonal/annual summaries of the raster stack values. When used, this automatically defines t_position = saddle.}

\item{t_position}{= A character vector of either NULL, "past", "saddle", or "future". If NULL, then values are extracted based on matching up the observation point with the dynamic raster stack at the level specified in the t_summ character vector (e.g., "daily", "monthly", "seasonal", "annual". If not, then summaries are calculated leading up to and including the observation time ("past"), saddled around around the observation time ("saddle"), or including and in the future of the observation time ("future").}

\item{sf_points}{= SF spatial points object specifying the locations where we want to extract raster stack values}

\item{date_col_name}{= The column name of sf_points with the date information, formatted as YYYY-MM-DD}

\item{df_sf}{= Character string of "df" or "sf" signaling whether the returned object should be a dataframe or sf object}
}
\value{
Either an SF object or dataframe, with the covariate values appended as new columns to the point locations file. This file is also saved in out_dir.
}
\description{
This function summarizes spatio-temporal 2d gridded data at sample point locations. The function takes as the primary inputs a stack of one dynamic variable and a sf spatial points object. The function then does the extraction by overlaying the spatial points onto the raster stack and gathering the value for each raster layer in the stack at all of the point locations. Then, depending on the t_summ and t_position arguments, the function summarizes the fully extracted data (all points all locations) at the appropriate temporal scale and position (leading up to, saddling, or in the future) relative to the observation time. The function returns either the original sf_points data set with a new column for the dynamic raster layer values or a dataframe of the same structure and also saves this file either as "model_covs.rds" if new_file_name is NULL or as the name provided in new_file_name.
}
