% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate_functions.R
\name{dynamic_2d_extract_wrapper}
\alias{dynamic_2d_extract_wrapper}
\title{Extract dynamic covariates at point locations wrapper}
\usage{
dynamic_2d_extract_wrapper(
  dynamic_covariates_list,
  t_summ,
  t_position,
  sf_points,
  date_col_name,
  df_sf,
  out_dir
)
}
\arguments{
\item{dynamic_covariates_list}{=}

\item{t_position}{}

\item{out_dir}{= Directory to save the combined dataframe as an .rds file}
}
\value{
A datafame with information of all unique tows with dynamic covariate values as new columns. This file is also saved in out_dir.
}
\description{
This short wrapper function is used to process all potential covariates in the dynamic_covariates_dir.
}
