% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combo_functions.R
\name{make_tidy_mod_data}
\alias{make_tidy_mod_data}
\title{Make a "tidy" model dataframe}
\usage{
make_tidy_mod_data(all_tows, all_tidy_occu, out_dir)
}
\arguments{
\item{all_tows}{= A dataframe with information on all unique tows, including any potential habitat covariates describing characteristics of these locations}

\item{all_tidy_occu}{= A dataframe with all occupancy records, created by `bind_nmfs_dfo_tidy_occu`}

\item{out_dir}{= Directory to save the tidy model dataframe as an .rds file}
}
\value{
A tidy model datafame with all the information (tows, habitat covariates, species occurrences) needed to fit a species distribution model. This file is also saved in out_dir.
}
\description{
This function brings together information on the unique tows, which will usually include measures of effort as well as habitat covariates, with the occupancy data to create a "tidy" model dataframe. The "tidy" model dataframe is set up so that each row is a unique tow - species - occupancy record with additional information for habitat covariates that we might include in a species distribution model.
}
