% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmfs_functions.R
\name{nmfs_get_tows}
\alias{nmfs_get_tows}
\title{NMFS tow data}
\usage{
nmfs_get_tows(nmfs_raw, out_dir)
}
\arguments{
\item{nmfs_raw}{Raw NMFS data, loaded with `nmfs_load`}

\item{out_dir}{Directory to save the cleaned dataframe as an .rds file}
}
\value{
A datafame with information of all unique NMFS tows. This file is also saved in out_dir.
}
\description{
This function reads in the nmfs_raw data (originally loaded as `survdat` and then renamed in the loading `nmfs_load` to `nmfs_raw`) and then creates a dataframe that has one row for every unique tow
}
