% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmfs_functions.R
\name{nmfs_make_tidy_occu}
\alias{nmfs_make_tidy_occu}
\title{NMFS "tidy" occupancy}
\usage{
nmfs_make_tidy_occu(nmfs_raw, nmfs_tows, species_table, out_dir)
}
\arguments{
\item{nmfs_raw}{= Raw NMFS data, loaded with `nmfs_load`}

\item{nmfs_tows}{=}

\item{species_table}{= Dataframe for the species of interest, including their DFO and NMFS species codes}

\item{out_dir}{= Directory to save the processed dataframe as an .rds file}
}
\value{
A "tidy" dataframe of species occurrences. This file is also saved in out_dir.
}
\description{
This function uses the raw NMFS data (`nmfs_raw`) to create a "tidy" occupancy dataset, where each row represents a unique tow - species - occupancy record for every species listed in `species_table`. In creating this "tidy" dataset, we impute "absence" observations, such that there is a record for every species in `species_table` at every unique tow.
}
