% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vast_functions.R
\name{process_vast_seasonal_data}
\alias{process_vast_seasonal_data}
\title{Make VAST dataset}
\usage{
process_vast_seasonal_data(tidy_mod_data, year_min, year_max, out_dir)
}
\arguments{
\item{out_dir}{}
}
\value{

}
\description{

}
