% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vast_functions.R
\name{read_cov_dat_csv}
\alias{read_cov_dat_csv}
\title{Read in covariate dataset}
\usage{
read_cov_dat_csv(cov_dat_path)
}
\arguments{
\item{cov_dat_path}{File path to sample data csv}
}
\value{
covariate data frame
}
\description{
A short function to read in covariate data given a file path to csv
}
