% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vast_functions.R
\name{read_polyshape}
\alias{read_polyshape}
\title{Read in shapefile}
\usage{
read_polyshape(polyshape_path)
}
\arguments{
\item{file_path}{File path to geospatial vector polygon file with .shp extension, specifying the location and shape of the area of interest.}

\item{factor_vars}{Names of factor columns that should be checked and converted if necessary}
}
\value{
SF poylgon
}
\description{
A short function to read in a shapefile given a file path
}
