% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vast_functions.R
\name{read_samp_dat_csv}
\alias{read_samp_dat_csv}
\title{Read in sample dataset}
\usage{
read_samp_dat_csv(samp_dat_path)
}
\arguments{
\item{samp_dat_path}{File path to sample data csv}
}
\value{
sample data frame
}
\description{
A short function to read in sample data given a file path to csv
}
