% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate_functions.R
\name{static_extract}
\alias{static_extract}
\title{Extract static covariates at point locations}
\usage{
static_extract(rast, cov_name, sf_points, date_col_name, df_sf)
}
\arguments{
\item{rast}{= A raster layer of one static covariate variable}

\item{cov_name}{= The column name to use for the extracted value for the static covariate}

\item{sf_points}{= SF spatial points object specifying the locations where we want to extract raster stack values}

\item{date_col_name}{= The column name of sf_points with the date information, formatted as YYYY-MM-DD}

\item{df_sf}{= Character string of "df" or "sf" signaling whether the returned object should be a dataframe or sf object}
}
\value{
Either an SF object or dataframe, with the covariate value appended as a new column to the point locations file. This file is also saved in out_dir.
}
\description{
This function reads a raster layer and a sf spatial points object. The function then does a simple extraction by overlaying the spatial points onto the raster layer and gathering the value for each raster layer in the stack at all of the point locations. The function then returns the original sf_points data set as either an sf object or dataframe with columns for the raster layer as a tibble and also saves this file as "model_covs.rds".
}
