% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate_functions.R
\name{static_extract_wrapper}
\alias{static_extract_wrapper}
\title{Static extract at point locations wrapper}
\usage{
static_extract_wrapper(
  static_covariates_list,
  sf_points,
  date_col_name,
  df_sf,
  out_dir
)
}
\arguments{
\item{static_covariates_list}{= List with static raster layers}

\item{out_dir}{= Directory to save the combined dataframe as an .rds file}

\item{all_tows}{= All tows dataframe, created by `bind_nmfs_dfo_tows()`}
}
\value{
A datafame with information of all unique tows and appended columns for each of the static habitat covariates. This file is also saved in out_dir.
}
\description{
This short wrapper function is used to process all potential covariates in the static_covariates_dir.
}
