% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vast_functions.R
\name{vast_build_sdm}
\alias{vast_build_sdm}
\title{Build VAST SDM}
\usage{
vast_build_sdm(
  settings,
  extrap_grid,
  samp_dat,
  cov_dat,
  X1_formula,
  X2_formula,
  X_contrasts,
  Xconfig_list
)
}
\arguments{
\item{settings}{A tagged list with the settings for the model, created with `vast_make_settings`.}

\item{samp_dat}{A data frame with the biomass sample data for each species at each tow.}

\item{cov_dat}{A data frame with the covariate data for each tow.}

\item{X1_formula}{A formula for the first linear predictor.}

\item{X2_formula}{A formula for the second linear predictor.}

\item{X_contrasts}{A tagged list specifying the contrasts to use for factor covariates in the model.}

\item{Xconfig_list}{A tagged list specifying the covariate effects for first and second linear predictors.}

\item{input_grid}{An extrapolation grid, created with `vast_make_extrap_grid`.}
}
\value{
A VAST `fit_model` object, with the inputs and built TMB object components.
}
\description{
Build VAST species distribution model, without running it. This can be helpful to check settings before running `vast_fit_sdm`. Additionally, it can be helpful for making subsequent modifications, particularly to mapping.
}
