% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vast_functions.R
\name{vast_fit_sdm}
\alias{vast_fit_sdm}
\title{Fit VAST SDM}
\usage{
vast_fit_sdm(vast_build_adjust)
}
\arguments{
\item{vast_build}{A VAST `fit_model` object.}
}
\value{
A VAST fit_model object, with the inputs and and outputs, including parameter estimates, extrapolation gid info, spatial list info, data info, and TMB info.
}
\description{
Fit VAST SDM
}
