% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vast_functions.R
\name{vast_make_adjustments}
\alias{vast_make_adjustments}
\title{Adjust VAST SDM}
\usage{
vast_make_adjustments(vast_build, adjustments = NULL)
}
\arguments{
\item{vast_build}{A VAST `fit_model` object.}

\item{adjustments}{Either NULL (default) or a tagged list identifying adjustments that should be made to the vast_build `fit_model` object. If NULL, the identical model defined by the `vast_build` is run and fitted.}
}
\value{
A VAST fit_model object, with the inputs and built TMB object components.
}
\description{
Make adjustments to VAST SDM and the model returned in `vast_build_sdm`. This can either be the exact same as the one built using `vast_build_sdm`, or it can update that model with adjustments provided in a tagged list.
}
