% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vast_functions.R
\name{vast_make_coveff}
\alias{vast_make_coveff}
\title{Make VAST covariate effect objects}
\usage{
vast_make_coveff(X1_coveff_vec, X2_coveff_vec)
}
\arguments{
\item{X1_coveff_vec}{A vector specifying the covariate effects for first linear predictor.}

\item{X2_coveff_vec}{A vector specifying the covariate effects for second linear predictor.}
}
\value{
A list with covariate effects for the first linear predictor (first list slot) and second linear predictor (second list slot).
}
\description{
Create covariate effects for both linear predictors
}
