% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vast_functions.R
\name{vast_make_extrap_grid}
\alias{vast_make_extrap_grid}
\title{Make VAST extrapolation grid settings from a shapefile}
\usage{
vast_make_extrap_grid(shapefile, cell_size)
}
\arguments{
\item{shapefile}{A geospatial vector sf polygon file, specifying the location and shape of the area of interest.}

\item{cell_size}{The size of grid in meters (since working in UTM). This will control the resolution of the extrapolation grid.}
}
\value{
Tagged list containing extrapolation grid settings needed to fit a VAST model of species occurrence.
}
\description{
Create a list of with information defining the extrapolation grid and used by subsequent VAST functions, leveraging code here: https://github.com/James-Thorson-NOAA/VAST/wiki/Creating-an-extrapolation-grid.
}
