% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vast_functions.R
\name{vast_make_sample_data}
\alias{vast_make_sample_data}
\title{Make VAST sample dataset}
\usage{
vast_make_sample_data(
  all_sample_dir,
  year_min,
  year_max,
  nmfs_species_code,
  out_dir
)
}
\arguments{
\item{out_dir}{}
}
\value{
Combined sample data frame
}
\description{
A short function that processes a sample dataset to work with VAST function and, if needed, extraction functions to get other covariate values before being used to fit VAST model
}
