% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vast_functions.R
\name{vast_make_settings}
\alias{vast_make_settings}
\title{Make VAST model settings}
\usage{
vast_make_settings(extrap_grid, FieldConfig, RhoConfig, bias.correct)
}
\arguments{
\item{extrap_grid}{User created extrapolation grid from vast_make_extrap_grid.}

\item{FieldConfig}{A vector defining the number of spatial (Omega) and spatio-temporal (Epsilon) factors to include in the model for each of the linear predictors. For each factor, possible values range from 0 (which effectively turns off a given factor), to the number of categories being modeled. If FieldConfig < number of categories, VAST estimates common factors and then loading matrices.}

\item{RhoConfig}{A vector defining the temporal structure of intercepts (Beta) and spatio-temporal (Epsilon) variation for each of the linear predictors. See `VAST::make_data` for options.}

\item{bias.correct}{Logical boolean determining if Epsilon bias-correction should be done.}
}
\value{
Tagged list containing settings needed to fit a VAST model of species occurrence.
}
\description{
Create a list of model settings needed to fit a VAST model for species occurrence, largely copied from VAST::make_settings
}
